+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-- LNC 438 fix bug no aetheryteshortcut by alydev
-- 3856 We Come in Peace bug fix by alydev
-- CT/GS/IS quest fixes by Kiarra
+- Unlock White Wolf Gate automatically when required by quest by alydev